backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
bcrypt==4.2.1
python-dotenv==1.0.1
joblib==1.4.2
requests==2.31.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn==1.7.2
reportlab==4.4.10
